Rhone - small edits to add and edit movies
Tweaks to add and edit movies to show the auditorium room names instead of id numbers on the add page to better show which room they are being added to.
Also, added auditorium room name to the view movies so that customers know which room the movie will be shown in.